fixing dev deploy
Issue with using react hooks outside react component (Not capitalized)
